Desktop/school/microservice/try3/llmMicroService/api-gateway/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
python-multipart